feat(checkout): rework calculation * use BigDecimal for monetary values * add quantity to ItemPrice * update pricing service to calculate total based on quantity and offers
Refs: BE-03

+++ backend/src/main/java/com/supermarket/service/ItemDto.java
@@ -1,15 +1,17 @@
 package com.supermarket.service;
 
+import java.math.BigDecimal;
+
 import com.supermarket.service.ItemDto;
 
 public class ItemDto {
     private long id;
     private String name;
-    private int price;
+    private BigDecimal price;
     private OfferRule offer;
     private int quantity;
 
-    public ItemDto(long id, String name, int price, OfferRule offer, int quantity) {
+    public ItemDto(long id, String name, BigDecimal price, OfferRule offer, int quantity) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -25,7 +27,7 @@ public class ItemDto {
         return name;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
